--- src/main/java/ParseCommands.java
@@ -1,5 +1,0 @@
-import java.io.IOException;
-
-interface ParseCommands {
-    String execute(TaskList tasklist) throws DukeException, IOException;
-}
