+++ src/main/java/duke/exception/EmptyListException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class EmptyListException extends DukeException {
+    public EmptyListException() {
+        super("Nee's list is empty! Add some tasks?");
+    }
+}
