+++ src/main/java/duke/exception/InvalidDateException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidDateException extends DukeException {
+    public InvalidDateException() {
+        super("Invalid date format, should be [dd/MM/yy] [HHmm]");
+    }
+}
