+++ src/main/java/duke/exception/InvalidTaskException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidTaskException extends DukeException {
+    public InvalidTaskException() {
+        super("Nee doesn't know what task you've done! \nAdd a description.");
+    }
+}
