+++ src/main/java/duke/exception/TaskNotFoundException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class TaskNotFoundException extends DukeException {
+    public TaskNotFoundException() {
+        super("Invalid task number, please try again.");
+    }
+}
