+++ src/main/java/duke/task/Todo.java
@@ -1,5 +1,7 @@
+package duke.task;
+
 public class Todo extends Task {
-    Todo(String taskName) {
+    public Todo(String taskName) {
         super(taskName);
     }
