+++ src/main/java/duke/util/Commands.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum Commands {
     TODO,
     EVENT,
