+++ src/main/java/duke/util/Ui.java
@@ -1,4 +1,5 @@
-import java.util.ArrayList;
+package duke.util;
+
 import java.util.Scanner;
 
 public class Ui {
